djangoBlog/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.4
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-formtools==2.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-redis==4.11.0
django-reversion==3.0.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.8
MarkupPy==1.14
MarkupSafe==1.1.1
mistune==0.8.4
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
PyMySQL==0.9.3
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
